Data rec/FileRescue/requirements.txt
pillow
shutil
Flask==2.2.2
Flask-SocketIO==5.3.0
eventlet==0.33.1